refactor: move peer calculation from ConnectionManager to PeerManager
Moves calculate_target_peers from ConnectionManager to PeerManager to properly
separate policy (peer targeting strategy) from mechanism (connection management).
PeerManager now calculates the target based on subnet count and passes it to
ConnectionManager via the renamed set_target_peers method.

This improves separation of concerns and eliminates the awkward pattern where
PeerManager had to call ConnectionManager::calculate_target_peers before
instantiating the ConnectionManager.
